+++ src/views/index.js
@@ -53,6 +53,8 @@ export { default as signature } from './signature/index.js';
 export { default as struct } from './struct/index.js';
 
 export { default as table } from './table/table.js';
+export { default as tableHeader } from './table/table-header.js';
+export { default as tableFooter } from './table/table-footer.js';
 export { default as tableRow } from './table/table-row.js';
 export { default as tableCell } from './table/table-cell.js';